[stockprice]: New stock price is now determined by weighted average (#251)
* [stockprice]: New stock price is now determined by weighted average

- Resync with proto

* Fixes possible price wraparound when calculation average price

* Added realAvgPrice float column to more accurately reflect the average price

+++ migrations/20170301153034_AddAvgLastPrice.up.sql
@@ -1,1 +1,1 @@
-ALTER TABLE Stocks ADD COLUMN avgLastPrice int(11);
+ALTER TABLE Stocks ADD COLUMN avgLastPrice bigint(11);
